+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - add basic container interface with implementation ([05dd46c](https://github.com/AVykhrystyuk/try-di/commit/05dd46c283afbb2b5152fa8626d034c7ce84d23d))
 - implement and test container.useFactory ([2ada2b8](https://github.com/AVykhrystyuk/try-di/commit/2ada2b8566cc22a599ef302a468004c1e590c725))
-- implement and test container.useValue ([75ede6e](https://github.com/AVykhrystyuk/try-di/commit/75ede6e957cbdf63b77a70e8b37c4e8ac474f2c9))
+- implement and test container.useClass ([75ede6e](https://github.com/AVykhrystyuk/try-di/commit/75ede6e957cbdf63b77a70e8b37c4e8ac474f2c9))
 - implement and test container.useValue ([242da9c](https://github.com/AVykhrystyuk/try-di/commit/242da9c1725774dff3d9d6ca81c0916e0de9016f))
 - implement container.tryVerifyAll ([eda6067](https://github.com/AVykhrystyuk/try-di/commit/eda606713e2fe2b2b8947210bd96b7d082ab79cc))
 - implement LegacyFactoryRegistry (for es5) ([400a419](https://github.com/AVykhrystyuk/try-di/commit/400a419d95084a8dece9aee988a4a50eb560a6bc))